done with history and notif burger menu and modal
