+++ client/public/index.html
@@ -26,7 +26,8 @@
     <title>OSU CS Applied Plan Portal</title>
   </head>
   <body>
-    <div id="portal"></div>
+    <div id="notification-portal"></div>
+    <div id="history-portal"></div>
     <div id="root"></div>
   </body>
 </html>
