+++ client/src/components/navbar/head_advisor_nav/HeadAdvisorNavDesktop.js
@@ -1,6 +1,6 @@
 import React from "react";
-import History from "../History";
-import Notifications from "../Notifications/Notifications";
+import History from "../history/History";
+import Notifications from "../notifications/Notifications";
 import {Link} from "react-router-dom";
 import UpdateCourses from "../UpdateCourses";
 import Logout from "../Logout";
