+++ client/src/components/navbar/head_advisor_nav/HeadAdvisorNavMobile.js
@@ -5,10 +5,8 @@ import PropTypes from "prop-types";
 import {slide as Menu} from "react-burger-menu";
 import {logout} from "../../../utils/authService";
 import ManageRoles from "../../manage_roles/ManageRoles";
-import History from "../History";
-import Notifications from "../Notifications/Notifications";
-import {useState} from "react";
-import MenuModal from "./MenuModal";
+import History from "../history/History";
+import Notifications from "../notifications/Notifications";
 
 function HeadAdvisorNavMobile({currentPlan}) {
   const styleMain = css`
@@ -70,7 +68,6 @@ function HeadAdvisorNavMobile({currentPlan}) {
 
   `;
 
-  const [isOpen, setIsOpen] = useState(false);
 
   const LogoutBtn = () => {
     return (
@@ -96,18 +93,6 @@ function HeadAdvisorNavMobile({currentPlan}) {
     );
   };
 
-  const NotificationBtn = () => {
-    return (
-      <div className="menu-item" onClick={handleOnOpenClick}>
-        <Notifications />
-      </div>
-    );
-  };
-
-  const handleOnOpenClick = () => {
-    setIsOpen(!isOpen);
-  };
-
   const HistoryBtn = () => {
     return (
       <div>
@@ -115,14 +100,13 @@ function HeadAdvisorNavMobile({currentPlan}) {
       </div>
     );
   };
-  console.log(isOpen);
+
   return (
     <div css={styleMain}>
-      <MenuModal isOpen={isOpen} handleClose={handleOnOpenClick} />
       <Menu right>
         <HomeBtn />
         <ManageRoleBtn />
-        <NotificationBtn />
+        <Notifications className="menu-item"/>
         <HistoryBtn />
         <LogoutBtn />
       </Menu>
